Create a new repo for user profile stuff. Move methods from current UserRepository to UserProfileRepository.
